The copying to/from slices was noticeably showing up in flamegraphs so took a shot at optimizing those and these changes seem to make a pretty good impact.  Due to the possibility of buffers being chained together, we can only do the optimization when there's enough contiguous storage but that is usually going to be the case.



before:

```
Timer precision: 50 ns
put_get_ux              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ get_ux_byte_aligned  64.81 ns      │ 94.05 ns      │ 71.41 ns      │ 73.36 ns      │ 100     │ 1000000
├─ get_ux_unaligned     306.3 ns      │ 477 ns        │ 341.5 ns      │ 357 ns        │ 100     │ 1000000
├─ put_ux_byte_aligned  119.1 ns      │ 159.5 ns      │ 127.5 ns      │ 129.7 ns      │ 100     │ 1000000
╰─ put_ux_unaligned     335 ns        │ 432.3 ns      │ 352.9 ns      │ 361.8 ns      │ 100     │ 1000000
```

after:

```
Timer precision: 39 ns
put_get_ux              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ get_ux_byte_aligned  68.62 ns      │ 110.8 ns      │ 77.04 ns      │ 78.86 ns      │ 100     │ 1000000
├─ get_ux_unaligned     92.38 ns      │ 128 ns        │ 103.2 ns      │ 105.9 ns      │ 100     │ 1000000
├─ put_ux_byte_aligned  123.5 ns      │ 169.7 ns      │ 133.9 ns      │ 135.2 ns      │ 100     │ 1000000
╰─ put_ux_unaligned     129.2 ns      │ 178.6 ns      │ 145.6 ns      │ 146.4 ns      │ 100     │ 1000000
```